Refactor resource managers to return their struct, not the interface
Returning an interface can reduce the code's flexibility by removing all
of the object's behavior that's not explicitly defined in the interface.
By returning a struct instead, we leave it up to the caller to decide
how the return should be used.

Expand on the networking manager's API to provide convenient means for
users to determine if ingresses and routes are available.

Signed-off-by: Lucas Caparelli <lucas.caparelli112@gmail.com>