Move default/validation logic to separate package
Previously logic for defaults and validation was distributed to each
resource manager. This makes it difficult to use default values without
creating import cycles and it complicates maintenance as there are
several sources of logic which need to be taken into account when making
changes.

If we keep all logic related to defaults and validation in a single
package we solve both of these issues.

This commit also makes sure the CR gets updated in the API server after
defaults were set and it has been validated, thus making sure the
desired state stored in the control plane is valid and can be easily
consulted at any time.

Fix #134