Implement identifiable logging
As the operator moves towards being cluster-scoped it's important that
users are able to quickly identify context from logs and can tell
exactly which Nexus CR is the subject of logs. To that effect this
change adds the "namespace" and "resource name" fields to structured
logging when applicable.

Additionally, standard output is now the default log output and sync
operations were removed as they're not necessary due to stdout's
unbuffered nature.

Signed-off-by: Lucas Caparelli <lucas.caparelli112@gmail.com>